chore: cleanup + comments + doc drift
Bulk cleanup batch from the last 6-lens sweep. No behaviour change.

Cleanup:
- Delete 5 unused intl keys (CERTIFICATION, DEGREE, LOADING, YEARS,
  YEARS_OF_EXPERIENCE) from both en-US.json and es-ES.json; parity
  verified.
- Remove unused `@react-router/cloudflare` dep — no import sites.
- Drop `export` on internal-only types: Cloudflare, DataTypes,
  ProjectsData, SkillHeatmapRow, RouteMetaOverrides, ForwardGroup.
  Delete unused `Project` type entirely.
- Extract prettyZodError → `app/data/format-zod-error.ts`
  (`formatZodError(source, err)`), deduping three verbatim copies
  across the schema files.
- Extract shared `passLoaderHeaders` → `app/utils/route-headers.ts`,
  replacing the identical 3-line `export function headers` in seven
  route files.
- Merge PERSON_JSONLD + WEBSITE_JSONLD into a single JSONLD_GRAPH
  under one `@context` — one `<script>` tag instead of two, saves a
  bit of gz.

Comment trim:
- workers/app.ts: drop the "CSP nonce" / "Security headers" /
  ".data rate limit" banner blocks. Replaced with 1-line pointers
  to docs/security.md.
- app/root.tsx: trim OG_IMAGE_URL/THEME_INIT_SCRIPT preambles,
  Layout narration about useRouteLoaderData/cspNonce/skip-link,
  and the JSX comments over the two inline scripts.
- app/data/{skills,education,projects}-schema.ts: drop the "Zod
  schema for X.json" file headers — filename says it.

Doc drift:
- AGENTS.md §3: `icons/` note drops "gitignored" (they're committed).
- AGENTS.md §5: add splat `*` route row.
- AGENTS.md §5: `/skills` cache header now shows `s-maxage=86400`.
- AGENTS.md §10: add pointer to the `privatizeCacheControl` invariant
  documented in docs/security.md.
- AGENTS.md §11: spec list adds projects/contact/error/a11y; visual
  coverage lists the real routes (/`, /education, /education/degree,
  /skills/1, /contact); Chromium-install note replaced with the
  Docker container reality.
- AGENTS.md §15 gotcha: rewrite the "1h cache" note to reflect the
  now-universal `Cache-Control: public, max-age=3600, s-maxage=86400`
  + `private` rewrite for SSR HTML.
- AGENTS.md §16 checklist: rewrite the `style.css` bullet from
  "links() spread" to the postcss-import `@import` pattern.
- wrangler.jsonc: RR v7 → RR v8 in headers; drop the Pages-migration
  narration.
- README.md: add pointer to docs/migrations/rr7-to-rr8.md.
- docs/security.md inline-script inventory: two JSON-LD blocks → one
  `@graph`.
