+++ app/data/education-schema.ts
@@ -1,12 +1,6 @@
 import { z } from 'zod';
 
-// Zod schema for public/data/education.json.
-//
-// Mirrors the skills-schema.ts pattern: single source of truth for
-// types + boot-time validation + path-precise errors on bad data.
-// Localizable string fields carry an optional `_es` sibling that
-// consumers resolve via `localized(item, key, locale)` in
-// app/utils/utils.tsx.
+import { formatZodError } from './format-zod-error';
 
 const yearMonth = z.string().regex(/^\d{4}-(0[1-9]|1[0-2])$/, 'Expected YYYY-MM');
 
@@ -67,31 +61,10 @@ const EducationSchema = z.object({
   certifications: z.array(certification),
 });
 
-// Format a Zod error tree into a multi-line, path-prefixed list for
-// log output — same shape as the skills-schema helper so failures in
-// either file produce a consistent, scannable error format.
-function prettyZodError(err: z.ZodError): string {
-  const lines = err.issues.map((issue) => {
-    const path = issue.path
-      .map((seg) => (typeof seg === 'number' ? `[${seg}]` : `.${String(seg)}`))
-      .join('')
-      .replace(/^\./, '');
-    return `  ✗ ${path || '(root)'}: ${issue.message}`;
-  });
-  return [
-    `education.json failed validation (${err.issues.length} issue${err.issues.length === 1 ? '' : 's'}):`,
-    ...lines,
-  ].join('\n');
-}
-
-// Parse + validate education.json content. Throws a pretty-printed
-// error that lists every failing path; fix all of them in one pass
-// instead of one-at-a-time. Called once per worker boot from the
-// route loaders that consume it.
 export function loadEducation(raw: unknown): z.infer<typeof EducationSchema> {
   const result = EducationSchema.safeParse(raw);
   if (!result.success) {
-    throw new Error(prettyZodError(result.error));
+    throw new Error(formatZodError('education.json', result.error));
   }
   return result.data;
 }
